+++ sources/Valkey.Glide/Abstract/IServer.cs
@@ -192,6 +192,25 @@ public partial interface IServer : IRedisAsync
     Task<DateTime> LastSaveAsync(CommandFlags flags = CommandFlags.None);
 
     /// <summary>
+    /// Explicitly request the database to persist the current state to disk.
+    /// </summary>
+    /// <seealso href="https://valkey.io/commands/bgsave/"/>
+    /// <seealso href="https://valkey.io/commands/bgrewriteaof/"/>
+    /// <seealso href="https://valkey.io/commands/save/"/>
+    /// <param name="type">The method of the save (e.g. background or foreground).</param>
+    /// <param name="flags">Command flags (currently not supported by GLIDE).</param>
+    /// <exception cref="NotImplementedException">Thrown if <paramref name="flags"/> is not <see cref="CommandFlags.None"/>.</exception>
+    /// <exception cref="NotSupportedException">Thrown if <paramref name="type"/> is not <see cref="SaveType.BackgroundSave"/>.</exception>
+    /// <remarks>
+    /// <example>
+    /// <code>
+    /// await server.SaveAsync(SaveType.BackgroundSave);
+    /// </code>
+    /// </example>
+    /// </remarks>
+    Task SaveAsync(SaveType type, CommandFlags flags = CommandFlags.None);
+
+    /// <summary>
     /// Return the current server time.
     /// </summary>
     /// <seealso href="https://valkey.io/commands/time"/>