+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support additional commands (#435):
+  - `BGSAVE`, `BGSAVE SCHEDULE`, and `BGSAVE CANCEL` (#436)
 - Custom socket address resolution support via callback (#392)
 
 ## 1.1.0